CPP05

+++ CPP05/ex00/Bureaucrat.cpp
@@ -2,12 +2,10 @@
 
 Bureaucrat::~Bureaucrat()
 {
-    // std::cout << "destructor called for bureaucrat" << std::endl;
 }
 
 Bureaucrat::Bureaucrat(std::string name, int grade) : name(name)
 {
-    // std::cout << "paramitrazed constructor called for bureaucrat" << std::endl;
     if (grade < 1)
         throw Bureaucrat::GradeTooHighException();
     else if (grade > 150)
@@ -17,13 +15,11 @@ Bureaucrat::Bureaucrat(std::string name, int grade) : name(name)
 
 Bureaucrat::Bureaucrat(const Bureaucrat& bureaucrat) : name(bureaucrat.name)
 {
-    // std::cout << "copy constructor called for bureaucrat" << std::endl;
     *this = bureaucrat;
 }
 
 Bureaucrat& Bureaucrat::operator=(const Bureaucrat& bureaucrat)
 {
-    // std::cout << "copy assignement operator called for bureaucrat" << std::endl;
     if (this == &bureaucrat)
         return (*this);
     this->grade = bureaucrat.grade;
@@ -68,6 +64,6 @@ const char* Bureaucrat::GradeTooLowException::what() const throw()
 
 std::ostream& operator<<(std::ostream& os, const Bureaucrat& bureaucrat)
 {
-    os << bureaucrat.getName() << ", bureaucrat grade " << bureaucrat.getGrade() << ".\n";
+    os << bureaucrat.getName() << ", bureaucrat grade " << (int)bureaucrat.getGrade() << ".\n";
     return (os);
 }
